Feature/cookies

See merge request akostyukov/todo-server!1

+++ app/main.py
@@ -1,16 +1,26 @@
 import cgi
 import re
+from http.cookies import SimpleCookie
 from http.server import BaseHTTPRequestHandler, HTTPServer
 
+from app.authServer import login_page, register_page, login, register, logout
 from app.config import env, session
 from app.forms import TaskForm
-from app.models import Task
+from app.models import Task, Token, User
+from app.response import Response
 
 
 class TaskHandler(BaseHTTPRequestHandler):
-    def set_headers(self, code, header, url):
+    def set_headers(self, code, header, url, token=None, cookie_expires=None):
         self.send_response(code)
         self.send_header(header, url)
+
+        if token:
+            cookie = SimpleCookie()
+            cookie['token'] = token
+            cookie['token']['expires'] = cookie_expires
+            self.send_header('Set-Cookie', cookie.output(header=''))
+
         self.end_headers()
 
     def do_GET(self):
@@ -27,55 +37,70 @@ class TaskHandler(BaseHTTPRequestHandler):
         self.handler('post', data)
 
     def handler(self, method, data=None):
-        response = routes(self.path, method, data)
+        response = routes(self.path, method, data, SimpleCookie(self.headers.get('Cookie')))
+
         self.set_headers(*response.headers)
         self.wfile.write(response.data.encode())
 
 
-class Response:
-    headers = ()
-    data = ''
+def task_list(cookie):
+    if cookie and Token.check_user(cookie):
+        headers = 200, 'Content-Type', 'text/html'
+        data = env.get_template('index.html').render(
+            form=TaskForm(),
+            tasks=session.query(Task).filter_by(status=True, user_id=User.get_user(cookie).id).all()[::-1],
+            done_tasks=session.query(Task).filter_by(status=False, user_id=User.get_user(cookie).id).all(),
+            current_user=User.get_user(cookie).login
+        )
 
-    def __init__(self, headers, data=''):
-        self.headers = headers
-        self.data = data
-
-
-def task_list():
-    headers = 200, 'Content-Type', 'text/html'
-    data = env.get_template('index.html').render(form=TaskForm(),
-                                                 tasks=session.query(Task).filter_by(status=True).all()[::-1],
-                                                 done_tasks=session.query(Task).filter_by(status=False).all()
-                                                 )
-    return Response(headers, data)
+        return Response(headers, data)
+    else:
+        headers = 302, 'Location', '/login'
+        return Response(headers)
 
 
-def add_task(data):
-    headers = 302, 'Location', '/'
+def add_task(data, cookie):
+    if cookie and Token.check_user(cookie):
+        headers = 302, 'Location', '/'
 
-    session.add(Task(data.getvalue('task')))
-    session.commit()
-
-    return Response(headers)
-
-
-def delete_task(task_id):
-    headers = 302, 'Location', '/'
-    session.query(Task).get(task_id).delete_task()
+        session.add(Task(data.getvalue('task'), User.get_user(cookie).id))
+        session.commit()
+    else:
+        headers = 302, 'Location', '/login'
 
     return Response(headers)
 
 
-def done_task(task_id):
-    headers = 302, 'Location', '/'
-    session.query(Task).get(task_id).set_done()
+def delete_task(task_id, cookie):
+    if cookie and Token.check_user(cookie):
+        headers = 302, 'Location', '/'
+
+        if User.get_user(cookie).id == session.query(Task).get(task_id).user_id:
+            session.query(Task).get(task_id).delete_task()
+    else:
+        headers = 302, 'Location', '/login'
 
     return Response(headers)
 
 
-def clear_all():
-    headers = 302, 'Location', '/'
-    Task.clear_all()
+def done_task(task_id, cookie):
+    if cookie and Token.check_user(cookie):
+        headers = 302, 'Location', '/'
+
+        if User.get_user(cookie).id == session.query(Task).get(task_id).user_id:
+            session.query(Task).get(task_id).set_done()
+    else:
+        headers = 302, 'Location', '/login'
+
+    return Response(headers)
+
+
+def clear_all(cookie):
+    if cookie and Token.check_user(cookie):
+        headers = 302, 'Location', '/'
+        Task.clear_all(cookie)
+    else:
+        headers = 302, 'Location', '/login'
 
     return Response(headers)
 
@@ -85,27 +110,32 @@ urls = [
     ('/', 'post', add_task),
     (r'/delete/(?P<task_id>\d+)', 'get', delete_task),
     (r'/done/(?P<task_id>\d+)', 'get', done_task),
-    ('/clear', 'get', clear_all)
+    ('/clear', 'get', clear_all),
+    ('/login', 'get', login_page),
+    ('/register', 'get', register_page),
+    ('/login', 'post', login),
+    ('/register', 'post', register),
+    ('/logout', 'get', logout),
 ]
 
 
-def routes(path, do_method, data):
+def routes(path, do_method, data, cookie):
     task_id = None
 
     for url, method, handler in urls:
         received_url = re.match(url, path)
 
-        try:
-            task_id = int(received_url.group('task_id'))
-        except:
-            pass
-
         if received_url is not None and received_url.group(0) == path and method == do_method:
+            try:
+                task_id = int(received_url.group('task_id'))
+            except:
+                pass
+
             if method == 'post':
-                return handler(data)
+                return handler(data, cookie)
             if task_id is not None:
-                return handler(task_id)
-            return handler()
+                return handler(task_id, cookie)
+            return handler(cookie)
 
 
 server = HTTPServer(('', 8000), TaskHandler)
